+++ packages/docusaurus/src/server/codegen/codegen.ts
